refactor: sections from `utils.core` -> `utils.xarray`
